A failed attempt to use flex for parsing json.
Bison isn't suited to parsing json, since it's bottom and our parse model is
more top down. Compare the events for [1, 2]:

Bottom Up:
 Int, Int, Array(Int, Int)

Top Down
 Open Array, Int, Int, Close Array.

So instead I tried to insert some glue to turn the lex tokens into parse
events. On monster.json this partially complete implementation runs in ~5ms,
which is in line with the previous hand coded imeplementation. I'm going to
stop going down this path but I'll check in the changes so we have it for the
future.